major event and threading clean up and overhaul
- full and better separation between main, draw/render and layout procs
- get rid of useless cmd indirection
- defer loading files until after initialization

initially tried to make all the threading stuff generic; going down
a rabbit-hole of channel implementations and portable threading,
screw it for now.
looked at tinycthread + tylertreat/chan, but the latter assumes
pthreads, needs to have all of its api renamed to match libthread
and fix the behavior of non-blocking functions, implement an alt(),
or add another layer between libthread and this.
no point for now.