linux: don't send a signal, terminate thread
why kill instead of communicate an exit?  because there's nothing to
clean up, we just want to interrupt the thread and have it die
immediately.

still arguable on linux; spawning threads here isn't as cheap, so we
could/should instead have one thread per graph and communicate
stop/restart/etc, then check for it in place of yield(); we'd send a
quit request, and wait for the thread to fully die before continuing.
we should do that, and make the threading generic, these details
shouldn't be reimplemented every time.