add new external memory indexing in linear time/space
combination of sort and awk; binarization cannot be done in awk,
so the final step is in python.
multipass method with about 8 steps that could maybe be reduced;
certainly a single c program with efficient sorting could take care
of this better, but this already works ok.
benchmarks for previous script on w520 with hapsim graphs:
	1e7 nodes: 27s
	1e8 nodes: 4m35
	1e9 nodes: 48m41
this is probably slower; harder to test because we need the
disk space for the stupid graph; might test with gzip, timing
only the indexing step.