fork awk, pass equivalent state and perform queries with it
awk doesn't have eval or a repl, which just fucks us.
the round about ways are cumbersome and not really beneficial.
the only ways to fix this are either to just forget this idea
and straight up integrate awk into strpg, or abandon awk for
python or whatever else could be depended on that isn't a
complete diarrhea fountain like python