strawk: plug eval leak
awk was never made for parsing new strings then discarding the
new parse tree. because of lots of assumptions and shortcuts
such as storing integer values as Node pointers (itonp et al),
freeing the tree node by node is not trivial. instead, allocate
nodes in slabs of 512 and free extra slabs once done; typically
evals will deal with short expressions so the first slab will be
reused over and over. there is still some bytes leaking from
strdup calls. strawk is now faster and consumes much less memory,
allowing loading graphs in the order of 1e6 nodes with ≤ 3gb of
memory.
this is a band-aid over a larger problem; this approach abuses
awk to the extreme, but a proper job would require rewriting
much of its grammar and code. since this will have to be done
later once the language itself is modified anyway, light
modifications are enough for now.