sokol_app: hack in event handler initialization after basic setup
in order to decouple input handling from rendering, input must
be initialized after x11 and sokol_gfx are set up; init is
called during the first frame and sokol_nuklear must be
initialized after sokol_gfx.
next, ideally we'd like to block until a new event occurs, but
then we have a race on quit because we can close the display
while trying to get new input, and segfault or hang forever on
exit. we also want to wait for events from the cmd pipe but we
use channels for that which would be annoying to change.
thus we have to use select() on the connection fd instead of
directly using XNextEvent() because the x11 api is [redacted].
we timeout after 500ms to poll for cmd events. perhaps we could
select on two fd's, a pipe from cmd instead, but it really
wouldn't change much.
all in all, this SUCKS but that's what you get for using linux.