sokol: stop drawing when there's nothing to redraw
this removes cpu load when idle, and reduces it when moving the
mouse around, etc.
the implementation reflects the landscape of the magical world
of linux; xlib is not thread-safe so we can't push a dummy event,
and select can mark an fd ready even though io would block.

based on ideas from https://github.com/floooh/sokol/pull/640