switch to 32bit identifiers
64bit identifiers waste a ton of space and complicate things
with the vague idea that we could have billion-node graphs;
however that's not what we're working with in practice.
it would be cheaper, when the time comes, to implement a
banking scheme, and even if the indirection sucks for some
of the nodes and edges, we'd still waste less space.
by the time we need even 40 bits, we'll have much larger
problems even with generating the graph, let alone
visualizing it. this does not justify shitting up
everything for an edge case.