graph: reject redundant flipped edges; draw: fix jumpy node angles and edge length fp truncation
circular mean calculation can also be approximated with:
	θ = atan2(∑sinθᵢ, ∑cosθᵢ)
either way, this still fails with exactly opposite angles
canceling each other, the result will be 0 in the current
implementation or nearly 0 in the previous, fp inaccuracies
causing crazy flipping around essentially a 0 angle.
not sure what the best way to fix it is yet.