fs/gfa: fix awk edge ids not corresponding to redges after load
redges are in node order after the first draw iteration;
the awk state for edges is currently only useful for printing
information.  it sucks, but keep the thread around and issue
the renames to awk after everything is loaded.  thus loading
time is unaffected and impact is practically none since it
will take a few seconds before the user will hover over edges
anyway.  other commands will be serialized automatically and
executed after the update is done with.