draw: fix issues due to mouse picking, improve implementation
duplicating/splitting rendering pipelines into two, one
rendering directly to screen, and one as before to an offscreen
buffer and also to a mouse picking buffer.  picking buffer
pipeline is used after a redraw has occurred or the view has
changed.  meanwhile, upon a mouse picking request, the contents
of the picking buffer are cached cpu-side, and updated only if
the buffer is stale and enough time has elapsed between the last
fetch and the mouse isn't being used to change the view (where
picking would be useless anyway).  fetching pixels is very slow,
one request can take 10ms or more, and since we pick any time
the mouse moves for the hover string, these fetches can line
up in a nasty way.  caching is particularly important on drag-
select, where we have to check a bunch of pixels.
this seems to improve reactivity in general, but it's difficult
to tell.