shift responsibility of book keeping to strawk
instead of a model with mixed and duplicated state between
visualizer and strawk, leave it all (most) to strawk.
the visualizer then only deals with nodes and edges without
any particular ordering or meaning; strawk will map objects
to nodes, edges, or components. this makes coarsening and
graph editing in general much simpler to handle. using
the addon() strawk extension, it can execute builtins
which have full access to the rest of the program. adding
onto this, we can get rid of a lot of back and forth
communication in cmd and simplify other things, for a net
gain in responsiveness and load times.

- awkext: implement nodecolor() as an addon
- awkext: more sanity checks; guard against racing
- awkext: fix quoting issues by parsing value strings
  ourselves; also helps reducing traffic somewhat
- awkext: don't let valbuf become too large, flushing it
  automatically after a threshold (tuned on x250)
- awkext: colors are given an alpha component if they
  are of the form #rrggbb
- awk: fix scheduling and order with multiple files;
  handle load profiling more intelligently
- draw: assume node length is final (for now) and
  re-scale once at startup, where we can also assume
  that all rnodes are primed and ready
- draw: fix issue where only half of the edges are
  rendered if graph is not redrawn
- fs: delegate most tag handling to strawk; delegate
  loadbatch calls as well; bigger gfa's now
  perceptually load much faster
- fs: fix issue where edge tags may remain unloaded
  in a batch
- graph: move length, color attributes outside of the
  environment; make nodes and edges global objects
  rather than properties of a graph, which aligns with
  rnodes and strawk
- layout: don't fully init layouting twice on startup;
  ensure that a sane initial state is still provided
  in case a layout file is imported
- layout: fix some of the scheduling issues on startup

highlighting is now broken but will be fixed by
implementing select()/deselect() as an addon.