sokol: don't draw from offscreen texture at all; merge shaders
we don't want multisampling when drawing to the picking texture,
but we can't simultaneously draw with multisampling to another
texture, so instead just don't do anything BUT drawing to the
picking texture, before doing the rest anyway. with msaa
enabled, this reduces performance somewhat, but is still much
smoother and avoids both all the complication and drawing
issues.

- up to date sokol-shdc output, consolidate edge/node shaders
  into the same module, avoiding redundant uniform declarations