look up visible edges manually; don't store in strawk at all
needed for coarsening. pending a better solution, add a
translation table from redges[], stored in arbitrary order,
to edges[], then look up left node in nodes[]. it sucks, but
we need to order nodes and edges better to avoid all of this,
or revisit how edges are stored and managed globally.
further, don't uselessly store strings in strawk; these can't
be exploited easily and will be invalidated by coarsening.
instead, manually pass a value to print when mouse picking.
there's a lot of redundancy there, but it's due to the same
problems.