make: use glcore on arm64, same as on amd64
gl es mouse picking code is broken and segfaults immediately.
this wasn't caught before because arm64 tests were not done
with it. if this runs on the switch, it would probably run
on rpi4+ and others. couldn't find how to fix this, so for
now we'll just use core opengl for arm64.