strawk: shrink Cell type and use ZV helper constant
tval doesn't need to be 4 bytes wide, and by moving it after
the two first byte-sized fields, we save 8 bytes total with
the struct padding (56 to 48 bytes).
we can't use ZV in initialization as it isn't a compile-time
constant, but we also can't #define ZV as (Value){.i = 0}
because that only works for initialization in kenc (for now?).