strawk: set temp cell type explicitely, rename literal $0
only set STR for a number once it gains a string conversion.
this avoids some easy breakages and simplifies a bit of code.
we want a number to be treated only as such until involved in
string operations or printing.
STR|NUM is set for preallocated variables to avoid string
conversion from the actual value, either as a shortcut or
because it would be different.
rename literal0 to avoid this same situation. $0 should be
STR|NUM but naming it 0 means that whenever we assign the
constant 0 to a number, we actually assign this cell and its
type; checks like if(var) then are actually equivalent to
if(var != ""), not if(var != 0).