strawk: add externally managed flat arrays
flat arrays are awk arrays that have attached (in c code) a
flat array of a certain type. the buffer is managed
externally, ie. external code retains its ownership.
strawk is not allowed to delete it, delete elements or
resize the array. in tests are disallowed for now, but
for-in loops loop through the array in order. since the
buffer is external, its value is re-fetched every time
it's accessed. the type of the array is set on attach
and should never be changed. if the buffer is realloced,
external code should take care to reattach the new
pointer properly. a regular array is set as an ids array
to permit accessing elements by index or string value.
this same ids array will be shared in strangepg across
all flat arrays.