strawk: don't error out if we can't assign to ptr value
errors are touchy, they can mess up the state,
but most of all, in operations like loadbatch they
can prevent any update.
instead, just skip updating the buffer.
the rest of the state is kept around to free cells
as usual.
set values are *not* turned into NAs or NaNs.