graph/coarse: improve ct construction performance
- get rid of the deg hash table, overload cnodes which is
  in any case unused during ct construction
- cache last child and top node in cnodes; even though this
  is mostly a problem with pathological cases which shouldn't
  happen much in practice, there's not much sense in
  recomputing these every time, and these are not fast
- use a flag to signal that the coarsening tree is ready
  instead of cnodes; not great but safer
- have to go back and properly regenerate cnodes/cedges after
  this is done.

the first iteration is always way faster, so getting rid
of the hash tables might be a good way to further increase
performance; we're also still storing mirrored edges and
our edge set is inflated as a result.
from benchmarking and timing, it seems as though the
hash tables themselves aren't particularly slow, but
at the 1e7+ scale the constant costs start to amount to
a lot.
fs/ctab needs to be updated separately.