cmd: fix not reinitializing fixed/initial pos across collapses
pos0 and flags are not populated or reset on collapse/expand,
and in fact still need to be, but this will be done once the
types are refactored. remove pos0 field and query tags when
layout is reset. since the visible set is supposed to be much
smaller than the entire collection, this can only benefit us.
also clean up layout code a bit.